Moving execute_interesting_file to Package class due to how packages are initialized
